requirements.txt
# Required Python packages for TCRMP 3D Processing Pipeline

# Core dependencies
pyyaml>=6.0.1       # For YAML configuration file parsing
pandas>=2.0.0       # For data manipulation and CSV handling
numpy>=1.24.0       # For numerical operations
opencv-python>=4.8.0  # For video frame extraction (step0.py)

# Optional dependencies
matplotlib>=3.7.0   # For visualization (if needed)
pillow>=10.0.0      # For image manipulation 